requirements.txt
dnspython==2.2.0
email-validator==1.1.3
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.4
